requirements.txt
datasets
pandas
numpy
mne
scipy
s3fs
pytz
tqdm
pyarrow
multiprocess
omegaconf
hydra-core
h5py
torch
pyyaml
pydantic
seaborn
matplotlib
scikit-learn
captum
einops
timm
wandb
mne-bids
comet-ml
braindecode
omegaconf